plugins/ocp: Adjusted variable type size carrying UUID index to 8 bit.
Changed uuidx type to __u8 in the related function calls,
because that's the type that will feed struct nvme_get_log_args.

+++ plugins/ocp/ocp-fw-activation-history.c
@@ -162,7 +162,7 @@ int ocp_fw_activation_history_log(int argc, char **argv, struct command *cmd,
 	if (err)
 		return err;
 
-	int uuid_index = 0;
+	__u8 uuid_index = 0;
 
 	/*
 	 * Best effort attempt at uuid. Otherwise, assume no index (i.e. 0)
